Rename package org.goobi.mq to org.kitodo.production.interfaces.activemq
