+++ Kitodo/src/main/java/org/kitodo/production/helper/Helper.java
@@ -36,11 +36,11 @@ import javax.faces.context.FacesContext;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.goobi.mq.WebServiceResult;
 import org.kitodo.production.enums.MessageLevel;
 import org.kitodo.production.enums.ReportLevel;
 import org.kitodo.production.helper.messages.Error;
 import org.kitodo.production.helper.messages.Message;
+import org.kitodo.production.interfaces.activemq.WebServiceResult;
 
 /**
  * Extends Helper from Kitodo Data Management module.
